Moved IEquatable and IComparable into MessageIdData partial class
